requirements.txt
langchain==1.2.13
langchain-core==1.2.20
langchain-community==0.4.1
langchain-anthropic==1.4.0
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.3

faiss-cpu==1.13.2
sentence-transformers==5.3.0
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-dotenv==1.2.2
pydantic-settings==2.13.1
